requirements.txt
streamlit
langchain
langchain-community
google-generativeai
python-dotenv

